refactor(provider): split useSendTransaction hook into focused modules
Reorganizes the monolithic useSendTransaction hook into separate, focused modules for better maintainability:

- use-send-transaction.ts: core transaction logic and state management
- use-send-transaction-display.ts: UI helpers, toast messages, and state conversion utilities
- use-send-transaction-setup.ts: wallet connection and metadata setup
- use-transaction-fee.ts: dedicated fee estimation logic

Updates playground page to use the new fee estimation hook. This is a pure refactoring with no functional changes - just improved separation of concerns between transaction logic and UI utilities.
